requirements.txt
pandas>=2.2            # core dataframe work
pyarrow>=17.0          # Parquet read/write
numpy>=1.26            # vector ops
scipy>=1.12            # required by scikit-learn
scikit-learn>=1.5      # TF-IDF, vectorizers
snscrape>=0.7.0        # Twitter/X scraping (no paid API)
vaderSentiment>=3.3    # sentiment scoring
emoji>=2.11            # emoji stripping
regex>=2024.5.15       # Unicode-safe regex for hashtags
matplotlib>=3.9        # lightweight plotting
pyyaml>=6.0            # config file parsing
tqdm>=4.66             # progress bars
